+++ components/ExamRenderer.tsx
@@ -2,9 +2,10 @@
 
 import { checkExam } from "@/app/actions"
 import { useRouter } from "next/navigation"
-import { FormEvent, ReactNode, useEffect, useState } from "react"
+import { FormEvent, useEffect, useState } from "react"
 import ReactMarkdown from "react-markdown"
 import remarkGfm from "remark-gfm"
+import { components } from "./ReactMarkdownComponents"
 
 type ExamRendererProps = {
   markdown: string | undefined
@@ -41,95 +42,14 @@ export default function ExamRenderer({ markdown }: ExamRendererProps) {
 
   return (
     <form onSubmit={handleSubmit} className="prose">
-      <ReactMarkdown
-        remarkPlugins={[remarkGfm]}
-        components={{
-          li({ children }) {
-            // Extraemos el texto plano del <li>
-            const raw = String(children)
-
-            // console.log({ children })
-            // console.log({ raw })
-
-            return formatReactMarkdown(raw, children)
-          },
-        }}
-      >
+      <ReactMarkdown remarkPlugins={[remarkGfm]} components={components}>
         {exam}
       </ReactMarkdown>
 
-      <input type="submit" />
+      <input
+        className="bg-light-black hover:bg-purple rounded-md p-2 text-white transition-colors"
+        type="submit"
+      />
     </form>
   )
-}
-
-function formatReactMarkdown(raw: string, children: ReactNode) {
-  //   console.log("typeof de children: ", typeof children)
-  console.log({ children })
-
-  // Caso: checkbox
-  if (raw.startsWith("[object Object]") || raw.startsWith("[x]")) {
-    const checked = raw.startsWith("[x]")
-    const label = raw
-      .replace(/\[object Object\],?\s*/g, "")
-      .replace(/^,?\s*/, "")
-      .trim()
-    return (
-      <li className="w-fit list-none">
-        <label className="flex items-center gap-2">
-          <input type="checkbox" defaultChecked={checked} />
-          {label}
-        </label>
-      </li>
-    )
-  }
-
-  // Caso: radio
-  if (raw.startsWith("( )") || raw.startsWith("(x)")) {
-    const checked = raw.startsWith("(x)")
-    const label = raw.replace("( )", "").replace("(x)", "").trim()
-    return (
-      <li className="w-fit list-none">
-        <label className="flex items-center gap-2">
-          <input type="radio" name="radio-group" defaultChecked={checked} />
-          {label}
-        </label>
-      </li>
-    )
-  }
-
-  // Caso: ShortAnswer
-  if (raw.includes("<ShortAnswer />")) {
-    if (!Array.isArray(children)) return
-    const sentence = children && (children[1].props.children as string)
-    return (
-      <li className="w-fit">
-        <p>{sentence}</p>
-        <textarea
-          rows={2}
-          className="w-full rounded border p-2"
-          placeholder="Redacta tu respuesta aquí..."
-        />
-      </li>
-    )
-  }
-
-  // Caso: Essay
-  if (raw.includes("<Essay />")) {
-    if (!Array.isArray(children)) return
-    const sentence = children && (children[1].props.children as string)
-    return (
-      <li className="w-fit">
-        <p>{sentence}</p>
-        <textarea
-          rows={6}
-          className="w-full rounded border p-2"
-          placeholder="Redacta tu respuesta aquí..."
-        />
-      </li>
-    )
-  }
-
-  // Fallback normal
-  return <li>{children}</li>
 }
